xfail issue2005 on Darwin/arm64

+++ testsuite/gna/issue2005/testsuite.sh
@@ -17,9 +17,13 @@ if c_compiler_is_available; then
 	libdir="$(dirname "$lib_path")"
 	incdir="$("$GHDL" --libghdl-include-dir)"
 
-	$CC -I"$incdir" test.c "$("$GHDL" --libghdl-library-path)"
-	$CC -I"$incdir" -L"$libdir" test.c -l"$lib"
-	rm -f a.out
+	if [ `uname -s` = Darwin -a `uname -m` = arm64 ]; then
+	    echo "XFAIL on macos arm64"
+	else
+	    $CC -I"$incdir" test.c "$("$GHDL" --libghdl-library-path)"
+	    $CC -I"$incdir" -L"$libdir" test.c -l"$lib"
+	    rm -f a.out
+	fi
 fi
 
 echo "Test successful"
